+++ src/modules/process/__init__.py
@@ -1,0 +1,7 @@
+"""
+Process Manager Module
+Advanced process management
+"""
+from .manager import ProcessManager
+
+__all__ = ['ProcessManager']